Added a memory controller and all mem calls go through it.  This should let programmers flip between banks easily.  This only allows the cpu to talk to one memory module at once.  This is very different from like a C64 that would cram several physical chips on the bus at once for like character rom, basic stuff, cartridge etc.
Made the stack pointer 16 bits.  256 is just too small for a stack.  So one whole memory module (up to 64k) is stack space.

Thinking about a mechanism to lock a memory module to a specific device while they are using it, but not sure what the cpu will do if it can't get what it wants.

With this paged stuff I should be able to have around 1.6 megs of ram.  I could always make the index a short for even more if needed, but that's plenty for doing like small textures or 2D tiles and whatnot.

Graphics modes will probably be like 320x180 for widescreen or 320x200 for squareish screens.  Would probably use a palette, but maybe could do one chip for red, green, blue?  Maybe I can put blits and polygon draws and such in the rom library that map to directx calls.